feat: canceled subscription should invoke invoice now
This helps with the user experience so that we don't have to wait for
the collector to complete.
